manage ingredient and page modif recipe

+++ backend/src/services/ingredient.service.ts
@@ -1,13 +1,14 @@
-import {Injectable} from "@nestjs/common";
-import {Observable} from "rxjs";
+import {ConflictException, Injectable} from "@nestjs/common";
+import {mergeMap, Observable} from "rxjs";
 import {IIngredientService} from "./iingredient.service";
 import {IngredientDao} from "../repositories/ingredient.dao";
 import {IngredientEntity} from "../entities/ingredient.entity";
+import {RecipeDao} from "../repositories/recipe.dao";
 
 @Injectable()
 export class IngredientService implements IIngredientService {
 
-    constructor(private ingredientDao: IngredientDao) {
+    constructor(private ingredientDao: IngredientDao, private recipeDao: RecipeDao) {
     }
 
     getAll(filters?: { [key: string]: any }): Observable<IngredientEntity[]> {
@@ -23,6 +24,14 @@ export class IngredientService implements IIngredientService {
     }
 
     delete(uuid: string): Observable<void> {
-        return this.ingredientDao.deleteIngredient(uuid);
+        return this.recipeDao.getAllRecipe().pipe(
+            mergeMap((recipe) => {
+                const ingredientUsed = recipe.some(recipe => recipe.amountIngredients.some(amount => amount.ingredient.uuid === uuid))
+                if (ingredientUsed) {
+                    throw new ConflictException('Ingredient is used in at least one recipe')
+                }
+                return this.ingredientDao.deleteIngredient(uuid);
+            })
+        )
     }
 }
